requirements.txt
# Core ML and Vector Database
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
FlagEmbedding>=1.2.0
pymilvus>=2.3.0

# Document Processing
langchain>=0.1.0
langchain-community>=0.1.0
langchain-huggingface>=0.1.0
langchain-docling>=0.1.0
docling>=1.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Progress and Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Optional: GPU support (uncomment if using CUDA)
# torch>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
